+++ src/BlockHorizons/BlockSniper/brush/types/LayerType.php
@@ -19,7 +19,7 @@ class LayerType extends BaseType{
 
 	public function __construct(Player $player, ChunkManager $level, \Generator $blocks = null){
 		parent::__construct($player, $level, $blocks);
-		$this->center = $player->getTargetBlock(100)->asVector3();
+		$this->center = $player->getTargetBlock($player->getViewDistance() * 16)->asVector3();
 	}
 
 	/**