+++ src/BlockHorizons/BlockSniper/commands/cloning/CloneCommand.php
@@ -32,7 +32,7 @@ class CloneCommand extends BaseCommand{
 			return;
 		}
 
-		$center = $sender->getTargetBlock(100);
+		$center = $sender->getTargetBlock($sender->getViewDistance() * 16);
 		if($center === null){
 			throw new InvalidBlockException("No valid block could be found when attempting to clone.");
 		}